+++ src/main/kotlin/io/openfuture/api/controller/api/ScaffoldApiController.kt
@@ -58,10 +58,12 @@ class ScaffoldApiController(
         return ScaffoldDto(scaffold)
     }
 
+    @PreAuthorize("hasRole('DEPLOY')")
     @GetMapping("/{address}/summary")
     fun getScaffoldSummary(@CurrentUser user: User, @PathVariable address: String): ScaffoldSummaryDto =
             service.getScaffoldSummary(address, user)
 
+    @PreAuthorize("hasRole('DEPLOY')")
     @PostMapping("/{address}/doDeactivate")
     fun deactivate(@CurrentUser user: User, @PathVariable address: String): ScaffoldSummaryDto =
             service.deactivate(address, user)
@@ -69,18 +71,21 @@ class ScaffoldApiController(
     @GetMapping("/quota")
     fun getQuota(@CurrentUser user: User): ScaffoldQuotaDto = service.getQuota(user)
 
+    @PreAuthorize("hasRole('DEPLOY')")
     @PostMapping("/{address}/holders")
     fun addShareHolder(@CurrentUser user: User, @Valid @RequestBody request: AddShareHolderRequest,
                        @PathVariable address: String) {
         service.addShareHolder(address, user, request)
     }
 
+    @PreAuthorize("hasRole('DEPLOY')")
     @PutMapping("/{address}/holders")
     fun updateShareHolder(@CurrentUser user: User, @Valid @RequestBody request: UpdateShareHolderRequest,
                           @PathVariable address: String) {
         service.updateShareHolder(address, user, request)
     }
 
+    @PreAuthorize("hasRole('DEPLOY')")
     @DeleteMapping("/{address}/holders")
     fun removeShareHolder(@CurrentUser user: User, @Valid @RequestBody request: RemoveShareHolderRequest,
                           @PathVariable address: String) {