+++ src/main/resources/db/migration/V1_3__scaffold_remove_field_user.sql
@@ -1,0 +1,1 @@
+ALTER TABLE scaffold_templates DROP COLUMN user_id;